Add and Style buttons
Added number buttons for calculations as well as operator buttons for
addition, subtraction, multiplication and division. Ensured that there
was adequate spacing between each of the buttons and also ensured that
the buttons are sized in a fitting manner.

Additionally changed the size of the input field for mathematical
expressions.